feat: adjust with common component
