+++ src/components/common/Input/index.tsx
@@ -7,10 +7,18 @@ interface InputProps {
   placeholder?: string;
   value?: string;
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  type?: string;
+  type?: 'text' | 'file';
+  refValue?: React.RefObject<HTMLInputElement>;
 }
 
-const SInput = ({ className = '', placeholder, value, onChange, type = 'text' }: InputProps) => {
+const SInput = ({
+  className = '',
+  placeholder,
+  value,
+  onChange,
+  type = 'text',
+  refValue
+}: InputProps) => {
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (onChange) {
       onChange(e);
@@ -19,6 +27,7 @@ const SInput = ({ className = '', placeholder, value, onChange, type = 'text' }:
 
   return (
     <input
+      ref={refValue}
       type={type}
       className={`px-4 py-2 border border-slate-300 rounded-lg bg-white focus:outline-none focus:ring-2 focus:ring-tree-300 ${className}`}
       placeholder={placeholder}
